requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
langchain==0.1.5
langchain-google-genai==0.0.8
langchain-community==0.0.17
pinecone-client==3.0.0
pypdf==6.5.0
python-multipart==0.0.6